messages: Added success messages for hospital
